# subscoping: Defer tag merging