--- a/scope_registry.go
+++ b/scope_registry.go
@@ -1,8 +1,28 @@
+// Copyright (c) 2020 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
 package tally
 
 import "sync"
 
-var scopeRegistryKey = KeyForPrefixedStringMap
+var scopeRegistryKey = keyForPrefixedStringMaps
 
 type scopeRegistry struct {
 	mu        sync.RWMutex
@@ -45,8 +65,7 @@ func (r *scopeRegistry) ForEachScope(f func(*scope)) {
 }
 
 func (r *scopeRegistry) Subscope(parent *scope, prefix string, tags map[string]string) *scope {
-	allTags := mergeRightTags(parent.tags, tags)
-	key := scopeRegistryKey(prefix, allTags)
+	key := scopeRegistryKey(prefix, parent.tags, tags)
 
 	r.mu.RLock()
 	if s, ok := r.lockedLookup(key); ok {
@@ -62,6 +81,7 @@ func (r *scopeRegistry) Subscope(parent *scope, prefix string, tags map[string]s
 		return s
 	}
 
+	allTags := mergeRightTags(parent.tags, tags)
 	subscope := &scope{
 		separator: parent.separator,
 		prefix:    prefix,