**Commits should be reviewed individually.**

This improves the performance of creating subscopes with new tags by
deferring merged map creation.

We currently create the merged map of tags first, and then calculate the
key which we use to determine if we have a cached subscope.

With this change, we calculate the key with slightly more complex logic
that inspects the unmerged maps and avoids paying the cost of merging
the maps if not needed.

To make this change, we first refactor scopeRegistry so that Scope
accesses it like a separate object.

---

I ran the benchmarks with `-count 3`.

**Before**

```
BenchmarkScopeTaggedCachedSubscopes-4             486406              2334 ns/op            1152 B/op          9 allocs/op
BenchmarkScopeTaggedCachedSubscopes-4             445890              2567 ns/op            1152 B/op          9 allocs/op
BenchmarkScopeTaggedCachedSubscopes-4             423003              2721 ns/op            1152 B/op          9 allocs/op
BenchmarkScopeTaggedNoCachedSubscopes-4           189488              6149 ns/op            2203 B/op         17 allocs/op
BenchmarkScopeTaggedNoCachedSubscopes-4           273436              6063 ns/op            2235 B/op         17 allocs/op
BenchmarkScopeTaggedNoCachedSubscopes-4           278360              5049 ns/op            2233 B/op         17 allocs/op
```

**After**

```
BenchmarkScopeTaggedCachedSubscopes-4             652670              1791 ns/op             832 B/op          8 allocs/op
BenchmarkScopeTaggedCachedSubscopes-4             669008              1953 ns/op             832 B/op          8 allocs/op
BenchmarkScopeTaggedCachedSubscopes-4             591230              1817 ns/op             832 B/op          8 allocs/op
BenchmarkScopeTaggedNoCachedSubscopes-4           234835              5876 ns/op            2269 B/op         18 allocs/op
BenchmarkScopeTaggedNoCachedSubscopes-4           315829              5058 ns/op            2237 B/op         18 allocs/op
BenchmarkScopeTaggedNoCachedSubscopes-4           243750              4993 ns/op            2264 B/op         18 allocs/op
```

On average,

- Cached hits are ~27% *faster*: ~687 ns/op
- Cache misses are ~7% *faster*: ~444 ns/op